backend/requirements.txt
Django==2.2.9
psycopg2==2.8.4
whitenoise==4.1.4
dj-database-url==0.5.0
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crispy-tailwind==0.3.0
cryptography==3.4.6
defusedxml==0.7.1
django-crispy-forms==1.11.1
django-extensions==3.1.1
django-filter==2.4.0
django-money==1.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-schema-graph==1.2.0
djangorestframework==3.12.2
fontawesome-free==5.15.3
idna==2.10
Markdown==3.3.4
oauthlib==3.1.0
pkg-resources==0.0.0
py-moneyed==0.8.0
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.3
uuid==1.30